+++ src/business/ScientificJournal.java
@@ -9,7 +9,7 @@ public class ScientificJournal extends Document {
 
     }
 
-    public ScientificJournal(int id, String name, String author, LocalDate publicationDate, int pageNumbers,
+    public ScientificJournal(long id, String name, String author, LocalDate publicationDate, int pageNumbers,
             String field) {
         super(id, name, author, publicationDate, pageNumbers);
         this.field = field;
